perf: code splitting on various places of the app
