+++ src/components/post/Reply.vue
@@ -52,7 +52,7 @@ import axios from 'axios'
 import Avatar from '@/components/Avatar.vue'
 import MoreIcon from '@/components/icons/More.vue'
 import BinIcon from '@/components/icons/Bin.vue'
-import { getPhotoFromIPFS } from '@/backend/photos'
+import { getPhotoFromIPFS } from '@/backend/getPhoto'
 import { getComment } from '@/backend/comment'
 import { createDefaultProfile, getProfile, Profile } from '@/backend/profile'
 import { getProfileFromSession } from '@/store/session'
