+++ src/layouts/bookmarks.vue
@@ -92,7 +92,7 @@ import Header from '@/components/Header.vue'
 import Footer from '@/components/Footer.vue'
 import UnauthPopup from '@/components/popups/UnauthPopup.vue'
 import { getProfile, Profile } from '@/backend/profile'
-import { getPhotoFromIPFS } from '@/backend/photos'
+import { getPhotoFromIPFS } from '@/backend/getPhoto'
 import { IPostResponse } from '@/backend/post'
 import { categories, IBackground, backgrounds } from '@/config'
 import { getBookmarksOfUser } from '@/backend/bookmarks'
